+++ components/layout/Header.styles.ts
@@ -1,0 +1,50 @@
+import styled from "styled-components";
+import Link from "next/link";
+import { StyledLink } from "@/styles/CommonStyles";
+
+const LoginLink = styled(StyledLink)``;
+
+const GlobalHeader = styled.header`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+const HeaderLeft = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+const HeaderLogo = styled(Link)`
+  margin-right: 16px;
+
+  @media (min-width: 768px) {
+    margin-right: 35px;
+  }
+
+  @media (min-width: 1280px) {
+    margin-right: 47px;
+  }
+`;
+
+const NavList = styled.ul`
+  display: flex;
+  list-style: none;
+  gap: 8px;
+  font-weight: bold;
+  font-size: 16px;
+  color: var(--gray-600);
+
+  @media (min-width: 768px) {
+    gap: 36px;
+    font-size: 18px;
+  }
+`;
+
+const NavItem = styled.li`
+  a:hover {
+    color: var(--blue);
+  }
+`;
+
+export { LoginLink, GlobalHeader, HeaderLeft, HeaderLogo, NavList, NavItem };
