+++ components/layout/Header.tsx
@@ -1,53 +1,15 @@
 import Logo from "@/public/images/logo/logo.svg";
-import Link from "next/link";
 import { useRouter } from "next/router";
-import styled from "styled-components";
-import { StyledLink } from "@/styles/CommonStyles";
+import Link from "next/link";
 
-const GlobalHeader = styled.header`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const HeaderLeft = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const HeaderLogo = styled(Link)`
-  margin-right: 16px;
-
-  @media (min-width: 768px) {
-    margin-right: 35px;
-  }
-
-  @media (min-width: 1280px) {
-    margin-right: 47px;
-  }
-`;
-
-const NavList = styled.ul`
-  display: flex;
-  list-style: none;
-  gap: 8px;
-  font-weight: bold;
-  font-size: 16px;
-  color: var(--gray-600);
-
-  @media (min-width: 768px) {
-    gap: 36px;
-    font-size: 18px;
-  }
-`;
-
-const NavItem = styled.li`
-  a:hover {
-    color: var(--blue);
-  }
-`;
-
-const LoginLink = styled(StyledLink)``;
+import {
+  GlobalHeader,
+  HeaderLeft,
+  HeaderLogo,
+  LoginLink,
+  NavItem,
+  NavList,
+} from "./Header.styles";
 
 function getLinkStyle(isActive: boolean) {
   return { color: isActive ? "var(--blue)" : undefined };
